IN-47 feat: move CreateNewPassword
