+++ app/auth/create-new-password/page.tsx
@@ -1,5 +1,5 @@
 import { Suspense } from 'react';
-import { CreateNewPasswordForm } from '@/src/features/auth/ui/reset-confirm/CreateNewPassword/CreateNewPasswordForm';
+import { CreateNewPasswordForm } from '@/src/features/auth/ui/forgot-password/CreateNewPassword/CreateNewPasswordForm';
 
 export default function Page({ searchParams }: { searchParams: Promise<{ [key: string]: string | undefined }> }) {
   return (
