+++ src/features/auth/ui/forgot-password/CreateNewPassword/CreateNewPasswordForm.module.scss
