+++ src/features/auth/ui/forgot-password/CreateNewPassword/CreateNewPasswordForm.tsx
@@ -1,5 +1,5 @@
 'use client';
-import styles from './CreateNewPasswordForm.module.scss';
+import styles from 'src/features/auth/ui/forgot-password/CreateNewPassword/CreateNewPasswordForm.module.scss';
 import { Input } from '@/src/shared/ui/Input/Input';
 import { Button } from '@/src/shared/ui/Button/Button';
 import Card from '@/src/shared/ui/Card/Card';
@@ -11,7 +11,10 @@ import { ROUTES } from '@/src/shared/config/routes';
 import { useAppDispatch } from '@/app/providers/store/hooks';
 import { setAppError } from '@/app/appSlice';
 import clsx from 'clsx';
-import { CreateNewPasswordFormValues, createNewPasswordSchema } from './createNewPasswordFormSchema';
+import {
+  CreateNewPasswordFormValues,
+  createNewPasswordSchema
+} from 'src/features/auth/ui/forgot-password/CreateNewPassword/createNewPasswordFormSchema';
 import { ErrorResponse } from '@/src/features/auth/lib/types/api.types';
 import { use } from 'react';
