+++ src/features/auth/ui/forgot-password/CreateNewPassword/index.ts
@@ -1,0 +1,1 @@
+export { CreateNewPasswordForm } from 'src/features/auth/ui/forgot-password/CreateNewPassword/CreateNewPasswordForm';
