--- src/features/auth/ui/reset-confirm/CreateNewPassword/index.ts
@@ -1,1 +1,0 @@
-export { CreateNewPasswordForm } from './CreateNewPasswordForm';
